1.0.0.33 Fix Import Application issues
Tests turned up a ton of issues with the Import-RedditApplication cmdlet.  Mostly typos but was also never updated to support Script Applications (UserCredential). IN addition, New-RedditApplication needed to have the UserCredential paramater made optional for the WebApp paramter set so that reconstituting the Reddit.Application objects could be more generalized.