Type Data Layout
Added Types Folder for Defining pstype data. It's a little awkward, but I think it is better than ps1xml's. connect-reddit.psm1 calls the Type files and ads the data.

Moved all the script properties of the Reddit.Applicationa nd Reddit.OAuthAccessToken objects out of the New- functions to be in alignment with other objects.

Also added a new property to Reddit.OauthAccessToken for the User Context the token is running under.